+++ src/libsodium/crypto_xof/shake128/ref/shake128_ref.h
@@ -15,16 +15,13 @@ typedef struct shake128_state_internal_ {
     unsigned char domain; /* Domain separation byte */
 } shake128_state_internal;
 
-int shake128_ref(unsigned char *out, size_t outlen, const unsigned char *in,
-                 unsigned long long inlen);
+int shake128_ref(unsigned char *out, size_t outlen, const unsigned char *in, size_t inlen);
 
 int shake128_ref_init(shake128_state_internal *state);
 
 int shake128_ref_init_with_domain(shake128_state_internal *state, unsigned char domain);
 
-int shake128_ref_update(shake128_state_internal *state,
-                        const unsigned char     *in,
-                        unsigned long long       inlen);
+int shake128_ref_update(shake128_state_internal *state, const unsigned char *in, size_t inlen);
 
 int shake128_ref_squeeze(shake128_state_internal *state, unsigned char *out, size_t outlen);
